platform/requirements.txt
accelerate==0.27.2
aiosmtplib==5.1.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
coverage==7.13.5
diffusers==0.27.2
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.0
greenlet==3.3.2
huggingface_hub==0.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
numpy<2
packaging==26.0
passlib==1.7.4
peft==0.12.0
Pillow==10.4.0
pluggy==1.6.0
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
psycopg2-binary==2.9.11
rembg==2.0.57
rsa==4.9.1
safetensors==0.4.2
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
torch==2.2.1
transformers==4.38.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
apscheduler==3.10.4